move randomString method to lib
