+++ apps/svelte/src/routes/password-generator/pin/+page.svelte
@@ -1,7 +1,5 @@
 <script lang="ts">
-	const lettersPattern = "abcdefghijklmnopqrstuvwxyzABCDEFGHIJKLMNOPQRSTUVWXYZ"
-	const numericPattern = "0123456789"
-	const symbolsPattern = ",._?-:]&*#~}$>(<)@^|{%!+="
+	import { randomString } from "lib/utils/string"
 
 	const letters = false
 	const numeric = true
@@ -10,37 +8,6 @@
 	let value = randomString({ length, letters, numeric, symbols })
 
 	$: value = randomString({ length, letters, numeric, symbols })
-
-	function randomString(props?: {
-		length?: number
-		letters?: boolean
-		numeric?: boolean
-		symbols?: boolean
-	}) {
-		const { length, letters, numeric, symbols } = {
-			length: 10,
-			letters: true,
-			numeric: false,
-			symbols: false,
-			...props,
-		}
-		const pattern = [
-			letters && lettersPattern,
-			numeric && numericPattern,
-			symbols && symbolsPattern,
-		]
-			.filter((v) => !!v)
-			.join("")
-		let value = ""
-		let index = length
-
-		while (index--) {
-			const charIndex = Math.floor(Math.random() * pattern.length)
-			value += pattern[charIndex]
-		}
-
-		return value
-	}
 </script>
 
 <h2 class="text-2xl font-semibold mb-6">PIN</h2>
