+++ apps/svelte/src/routes/password-generator/random/+page.svelte
@@ -1,7 +1,5 @@
 <script lang="ts">
-	const lettersPattern = "abcdefghijklmnopqrstuvwxyzABCDEFGHIJKLMNOPQRSTUVWXYZ"
-	const numericPattern = "0123456789"
-	const symbolsPattern = ",._?-:]&*#~}$>(<)@^|{%!+="
+	import { randomString } from "lib/utils/string"
 
 	let value = randomString()
 	let length = 20
@@ -9,37 +7,6 @@
 	let symbols = false
 
 	$: value = randomString({ length, numeric, symbols })
-
-	function randomString(props?: {
-		length?: number
-		letters?: boolean
-		numeric?: boolean
-		symbols?: boolean
-	}) {
-		const { length, letters, numeric, symbols } = {
-			length: 10,
-			letters: true,
-			numeric: false,
-			symbols: false,
-			...props,
-		}
-		const pattern = [
-			letters && lettersPattern,
-			numeric && numericPattern,
-			symbols && symbolsPattern,
-		]
-			.filter((v) => !!v)
-			.join("")
-		let value = ""
-		let index = length
-
-		while (index--) {
-			const charIndex = Math.floor(Math.random() * pattern.length)
-			value += pattern[charIndex]
-		}
-
-		return value
-	}
 </script>
 
 <h2 class="text-2xl font-semibold mb-6">Random Password</h2>
