+++ packages/lib/utils/string.ts
@@ -1,3 +1,7 @@
+const lettersPattern = 'abcdefghijklmnopqrstuvwxyzABCDEFGHIJKLMNOPQRSTUVWXYZ'
+const numericPattern = '0123456789'
+const symbolsPattern = ',._?-:]&*#~}$>(<)@^|{%!+='
+
 export function classNames (...args: (string | { [key: string]: boolean })[]) {
   const classes: string[] = []
   for (const arg of args) {
@@ -27,3 +31,34 @@ export const uniqueId = (() => {
   let id = -1
   return () => '' + ++id
 })()
+
+export function randomString (props?: {
+  length?: number;
+  letters?: boolean;
+  numeric?: boolean;
+  symbols?: boolean;
+}) {
+  const { length, letters, numeric, symbols } = {
+    length: 10,
+    letters: true,
+    numeric: false,
+    symbols: false,
+    ...props,
+  }
+  const pattern = [
+    letters && lettersPattern,
+    numeric && numericPattern,
+    symbols && symbolsPattern,
+  ]
+    .filter((v) => !!v)
+    .join('')
+  let value = ''
+  let index = length
+
+  while (index--) {
+    const charIndex = Math.floor(Math.random() * pattern.length)
+    value += pattern[charIndex]
+  }
+
+  return value
+}
